Reorganize Docker CI to run faster.
This restructures our Docker CI so that we do the builds for PR CI
checks in parallel where possible, and each platform shows up as it’s
own build.

This makes the PR CI checks run _much_ faster and also ensures that when
a given image build fails, it0s obvious which one failed.