Update pfsense.md
Change package links to newest versions.

The proposed change is because you can't find the new packages if you search the name of the old ones as py37 seems not to be supported any more. Likewise pfsense is using freebsd 12 not 11 anymore so if the use did find the correct package they might not realize the link needed would be pointing to the new version.