Update ansible.md
Fixed typos. 

However, the document needs some refinement (adding to my backlog). 
The document contains steps in form of paragraphs, instead of using ordered lists. 
It is also a bit generic at times, for example lines 49-51. 
"First, download the
[playbook](https://github.com/netdata/community/tree/main/netdata-agent-deployment/ansible-quickstart), move it to the
current directory, and remove the rest of the cloned repository, as it's not required for using the Ansible playbook."

What does 'current' mean in this context? Which file should be edited and which ones deleted? This will be explained later in the text, but it's not obvious from the beginning.