Send ML feature information with UpdateNodeInfo.
We achieve this by adding the `ml_{capable,enabled}` fields in
`system_info`. When streaming, these fields allow a parent to understand if
the child has ML and if it runs ML for itself.

The UpdateNodeInfo includes this information about a child, plus a
boolean that is set to true when the parent runs ML for the child.