Update packaging CI to only run a limited set of jobs on PRs.
The jobs for architectures other than 64-bit x86 consistently show no
issues unless something is broken for that architecture (which also
consistently breaks the static build for the same architecture) or that
distribution 9which also consistently breaks the 64-bit x86 build for
that distribution), so they functionally serve no real purpose when
being run on PRs.

This commit skips running those jobs on PRs only so that we do not waste
resources on them. They can be explicitly run by placing the exact words
`actions run all ci` in the PR description directly preceeded by a `/`.

This also enables running them on merges to master, which we previously
did not do, so that we can still confirm for certain that everything is
working.