Don't publish nightlies if they are unchanged since the last nightly.
This skips publishing nightlies if the last commit that modified
`packaging/version` is also the latest commit on the master branch _and_
the version listed in `packaging/version` ends in `-nightly`.

If both conditions are met, then there have been no changes since the
last nightly build, thus we don0t need to publish a new nightly build.

This check is limited to runs that occur as part of the `schedule`
trigger, to ensure that manually triggered builds can be still be done
without having to worry about changes.