Use one transaction to initialize dimension metadata
Prior to this PR, we'd execute multiple transactions for each newly
created dimension. This PR bundles each SQL statement we need to
execute for a new dimension in a single transaction. To support
this workflow, we wrap the collect/query ops and lazily initialize
the metadata of dimensions in one go the first time a dimension is
accessed.