PRs template: Include user information section
##### Summary
<!--
Describe the change in summary section, including rationale and design decisions.
Include "Fixes #nnn" if you are fixing an existing issue.
-->
As @thiagoftsm and @Ferroin suggested: 
We include a list of our PRs in the release notes. However, when you click on a PR as a user, who isn't involved in the day to day business, you might not understand what the PR does. This PR introduces a new section within "Additional information" to provide additional context for our users. 


##### Test Plan

<!--
Provide enough detail so that your reviewer can understand which test cases you
have covered, and recreate them if necessary. If our CI covers sufficient tests, then state which tests cover the change.
-->
N/A

##### Additional Information
<!-- This is usually used to help others understand your
motivation behind this change. A step-by-step reproduction of the problem is
helpful if there is no related issue. -->
The `details` tag renders a dropdown section and the `summary` tag is used to provide a custom heading. 
Please review this PR as source code and not the rendered version. The `details` tag contains a comment with guiding questions on how to write a good `For users` section.

<details> <summary>For users: How does this change affect me?</summary>
  <!--
Describe the PR affects users: 
- Which area of Netdata is affected by the change?
- Can they see the change or is it an under the hood? If they can see it, where?
- How is the user impacted by the change? 
- What are there any benefits of the change? 
-->
</details>