Treat auto-updates as a tristate internally.
If no options are specified, use the value `default` to indicate that
the default behavior is requested. Long-term, this will allow us to
easily have different defaults depending on other installation
parameters.

Shorter term, it allows us to explicitly indicate in telemetry events
whether the user asked for auto updates, asked for no auto updates, or
just went with the default.