Use explicit typecast to enum rrdset_flags
specifications say
An enumerator can be promoted to an integer value. However,
converting an integer to an enumerator requires an explicit
cast, and the results are not defined.

Therefore The bitwise OR operation you are performing results
in an int, which you then attempt to assign to a variable of
type rrdset_flags without a cast.

Fixes
| ml/Host.cc:167:9: error: assigning to 'RRDSET_FLAGS' (aka 'rrdset_flags') from incompatible type 'int'
|         rrdset_flag_set(RS, RRDSET_FLAG_ANOMALY_DETECTION);
|         ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~

Upstream-Status: Submitted [https://github.com/netdata/netdata/pull/12750]

Signed-off-by: Khem Raj <raj.khem@gmail.com>