Add support for detectiong cloud instance info using cloud-init.
This gives us immediate support for a wide variety of additional cloud
providers in our system info detection, and also allows us to provide
more information on certain providers.

Adds three new variables to the script output:
- NETDATA_INSTANCE_CLOUD_PLATFORM: Identifies the particular hosting
  platform in use. For example, this differentiates between EC2 and
  Lightsail on AWS.
- NETDATA_INSTANCE_CLOUD_INSTANCE_ZONE: Identifies the availability zone
  information for the domain, if we can determine it.
- NETDATA_INSTANCE_CLOUD_DETECTION: Identifies how the cloud instance
  info was obtained. Will either be `cloud-init`, or identify the
  particular IMDS that was queried.

This also makes a minor change in semantics for the
NETDATA_INSTANCE_CLOUD_PROVIDER variable, which will now be all
lowercase (this is the format returned by `cloud-init`, and thus is
needed for consistency).