Explicitly skip uploads and notifications in third-party repositories.
This changes how we handle uploads and Slack notifications in our
release build workflows so that they are skipped when running in
repositories other than our own.

This simplifies testing and verification of the core release code by
enabling testers to run a mock release build or mock nightly build to
completion without worrying about the uploads and slack notifications
failing the whole workflow.