Add a script to handle installing Netdata as a system service.
This uses the files installed as a result of the previous commit, and
provides more robust detection and system handling than the existing
code used for this purpose.

A subsequent commit will convert the various installation mechanisms to
use this script instead of their own internal code for this purpose.