Split monit and watchdog to their own groups.
Monit is used for a lot more than just network monitoring (it’s also a
service manager, and can even be used as an init system).

‘watchdog’ is a super-generic name, and while there may be some NMS with
that name, it also happens to match a local watchdog device manager (not
unusual on a server), so it really should be it’s own thing.