Add fallback to detected but not running service managers.
Currently supported for systemd and OpenRC. If no other service manager
is detected, but we detected that either systemd or OpenRC is present
but not running, fall back to them instead of outright failing.

This does not allow starting/stopping the agent using the service manager,
but _does_ happen to work for installing the service properly on an
offline system.