Add configuration option to drop logs if circular buffer is full.
Give the choice to the user about what will happen if the circular
buffer is full and there are pending logs. By default, the space
request function will block, which will result in delayed log
collection for the pending logs, but logs won't be lost.

Otherwise, if "circular buffer drop logs if full = 1", pending
logs can and will be dropped (but there will be no log collection
backlog for pending logs).