Rework logs query API in preparation for functions support.
Query API code is simplified and refactored, so that results
buffer is ready to support the functions contract. Also,
the results in data array now are returned as a JSON array,
to be easier for the frontend to parse them. The old results
format is still supported by providing the "data_format=
newline" argument to the GET parameters.