Fix the concurrency group name for PR labeler workflow. (#14460)
Instead of flagging on the git ref for the run, flag on the repository
ID and PR number. This ensures that previous runs get cancelled only
for the correct PR when a new run is triggered.

This should fix the seemingly random cases of PR labeler runs being
cancelled for no apparent reason on our PRs.