Add an option to the kickstart script to override distro detection.
Takes a single argument consisting of the values of the `ID`,
`VERSION_ID`, and `VERSION_CODENAME` fields from the os-release file
for the target distribution, separated by `:`. The codename part is
optional for non-DEB distros, in which case the final `:` may be left
off.

This allows for better handling of platforms that are ABI-compatible
derivatives of our supported platforms, both in that it allows users to
install with native packages on them, and in that it allows us to better
test such setups.