Replace `info` macro with a less generic name
The `info` macro is terribly generic, and can cause compilation problems
with other software.

In particular, newer versions of Protobuf (22+) pull in Abseil, which
defines an `info` function[^1] that is broken by the macro defined in
`libnetdata/log/log.h`.

This change will help unblock us at Homebrew, where we are trying to
switch our packages to using a newer version of `protobuf`.[^2]

[^1]: https://github.com/abseil/abseil-cpp/blob/e6c09ae4b2acd421a29706f86e66eaa422262ad0/absl/strings/internal/cordz_update_scope.h#L61
[^2]: https://github.com/Homebrew/homebrew-core/pull/133746