Fix package builds on Rocky Linux.
When switching from Alma Linux to Rocky Linux for our RHEL package
builds, we missed one place in our RPM spec file where a conditional was
checking whether we were building on Alma or not.

This fixes that conditional so it also checks for Rocky, which otherwise
fails the build in the same way that Alma would without the conditional.

It also future-proofs things so that we behave correctly on RHEL 10 when
that finally comes out.