Allows excluding v2 dashboard from installation
For certain Linux distributions, it's important that they ship only
software with free license. To help with that, add a flag which disable
installing the v2 dashboard.

The flag excludes the web/gui/v2/ directory from building, as well as
the index.html which downloads the latest v2 dashboard from server.
index.html of v1 will get installed instead, which means users will
seamlessly use the v1 dashboard in this case.

Bug: https://github.com/netdata/netdata/issues/16035