Added mkdir to the OpenSSL script to prevent errors
For default install on Ubuntu, the ssl subdirectory is not created, so the openssl command fails.  This addes the mkdir command, with -p so that it will not cause an error if the directory already exists.  This should solve the problem for Ubuntu and be harmless in any other case.