Disable local Go builds in our RPM packages.
We’re in the process of migrating the Go collectors to be part of the
main repo, and some of the intermediate steps don’t pay nice with the
current code in the spec file for handling of the Go plugin, so stub it
out for now.

Long-term this code is going to be removed completely (and some special
stuff will be done to handle the dependencies correctly since we need
a very recent version of Go that is not included on _most_ platforms),
as the Go code will be handled automatically as part of the CMake build
process, but for now it’s in the way and it’s simpler to just stub
it out.