Do not cancel detection thread.
An atomic flag is used before every iteration
of the detection thread. The flag is set right
before the cancellation request. It is possible
for the detection thread to have already exited
prior to sending the cancellation request.

The training thread most probably suffers from
the same issue, a separate PR will be opened
to fix it.