Move contrib/debian to the repo root.
This is where essentially all Debian tooling actually expects it to be
found, and it’s not been ‘contrib’ code for a very long time now so
keeping it in contrib makes no sense.

As this is the last thing that’s actually _in_ contrib, also remove the
contrib directory.