Correctly mark protobuf as required in find_package. (#17021)
At the point at which we call find_package() to locate protobuf, we know
that we do, in fact, need it for this build. Given this, we should be
marking it as required in the find_package() call so that if it’s not
found the configuration phase fails instead of things breaking in
potentially strange ways later on during the build itself.