Add plug to provide access to `/@{PROC}/@{pid}/limits`
This is the only plug that provides wide read access to other processes'
properties, opening up proper metrics on file descriptors and limits.

However, I don't think this is proper and `snapd` should expose another
way to get this type of access, because it isn't necessarily beholden to
using Docker.