Switch to using cmake directly.
This build takes cues from `contrib/debian` and also directly adjusts
the prefix to `/snap/netdata/current` and adjusts the default config to
only the needed changes to point to `/var/snap/netdata/common` for
config and `/current` for everything that goes into `/var`.

Additionally, a lot of runtime dependencies are taken from
snapcraft hints.