Generate header for /netdata.conf endpoint at build time.
This modifies how we are handling generation of the config file exposed
via the /netdata.conf endpoint and the dumpconfig command so that the
file header is provided by system/netdata.conf.

These changes achieve this by adding a CMake script to generate a header
file with a single define containing the text of system/netdata.conf.

CMake was chosen for this instead of something like Python because it’s
implicitly portable to all of our target systems, does not add a new
build-time dependency, and surprisingly enough it’s both simpler and
faster for this specific task.

Overall, this reduces duplication in the code, and ensures that users
get the same info in the comments at the top of netdata.conf regardless
of the source of the file.