Improve anacron detection in updater.
Instead of only checking the parent and grandparent processes, walk up
the process tree until we either find anacron or find init (or hit the
recursion limit for the function, which is currently 50 iterations).
This is needed because at least on some systems there is more than one
intermediate process between the script and anacron.