Handle SUID and file capabilities using CMake.
- Define the required permissions in CMake.
- Use CMake install hooks to apply them if not doing a packaging build.
- Use CMake to generate the required postinstall scripts if doing a
  packaging build.

The general idea here is to ensure that _all_ special permissions are
handled via CMake, so that we don’t need to change things in half a
dozen places each time we need to update what file capabilities or SUID
state we want on a given binary.

On Linux, this will use file capabilities if requested to do so, and
fall back gracefully to SUID if it cannot do so.

On macOS and FreeBSD, it will just use SUID (but we may eventually be
able to update the FreeBSD code to use only the required privileges).

On Windows it does nothing, because the concepts involved don’t really
exist there in a way we can usefully utilize.

Almost everything of interest here is in CMakeLists.txt and the newly
added files. The other changes are just cleaning things up so that the
new code actually gets used.