fix(cgroups.plugin): handle containers no env vars
When building container images "manually" (i.e., not with "docker
build"), it is possible to create them with no environment variables
specified. When containers are run with these images, they will have a
non-existent `.Config.Env`.

This causes no problems for `docker_like_get_name_command()`, however,
`docker_like_get_name_api()` will fail with:

> jq: error (at <stdin>:1): Cannot iterate over null (null)

When running netdata inside docker (or on a host that doesn't have
docker/podman binaries available), the end result is that it is not able
to resolve the cgroup IDs of those containers to their friendly display
names.