Add support for systemd NOTIFY_SOCKET in agent core.
This adds support for using systemd’s service notification functionality
to inform the service manager that Netdata has successfully finished
startup and that it is shutting down when told to do so.

The primary advantage of this for our usage is that it ensures that on
systemd systems Netdata will only be shown as running once it has
actually finished startup, instead of the current arrangement where it
is considered running as soon as the service manager forks the process.

This, in turn, gets a couple of other improvements:

- `systemctl start netdata` will only return once the agent is actually
  running.
- `systemctl start netdata` will properly report an error if there is a
  startup issue in the agent.
- Any services that depend on Netdata will only be started _after_ the
  agent finishes startup, ensuring that they can depend on the agent
  actually being running.

Internally, this uses a stand-alone implementation of this functionality
instead of relying on the one provided by libsystemd, as handling it
this way enables us to support it even on static builds. The actual
implementation is almost entirely copied from the MIT-0 licensed example
implementation provided by systemd itself.