Preliminary integration of the new otel-collector into the build.
Adding `-DENABLE_PLUGIN_OTEL=On` to CMake options will enable building
the plugin and installing it.

Currently disabled by default, and does not include packaging
integration yet.

The plugin itself can be built independently of the primary build system
but using the same build infrastructure that is used to build it as
part of the regular build by using CMake in the src/go/otel-collector
directory.