Add Docker tags for the last few nightly builds. (#19734)
This adds new Docker tags for our nightly builds to track prior nightly
builds relative to the current one.

The new tags take the form `edge-x` where the `x` is a number that
represents how many builds previous to the current build the tag tracks.
For example, `edge-1` is the build just prior to the current nightly
version, while `edge-3` is three builds prior to the current nightly
build.

These tags are rotated every time a nightly build is published. This
approach was chosen to avoid the significant complexity that would be
required to properly clean up old nightly build tags if we were tagging
by actual nightly version instead of tagging relative to the current
build.

The current configuration will track the 4 most recent nightly builds.

The special tag `edge-0` is a new alias for the current nightly build.
This is only needed to allow the tag update logic to work correctly, as
the workflow is intentionally configured to publish the new nightly
images before rotating the tags for the old ones (so that if rotation
fails it does not prevent us from publishing the current nightly).