Document switching from a native package to a static build (#20075)
* Document switching from a native package to a static build

Even though we document that static builds can be used on platforms that
are not, or no longer, supported, we do not explain how to _switch_ from
a native package to a static build. This PR addresses that.

* Link to start-stop documentation.

Co-authored-by: Fotis Voutsas <fotis@netdata.cloud>

* Add whitespace after code blocks

* Remove systemd commands, except step 4

This is because without step 4, the service will not work properly after
install.

* Apply suggestions from code review

* Apply suggestions from code review

---------

Co-authored-by: Fotis Voutsas <fotis@netdata.cloud>