Use parallelization for LTO processing to reduce build times.
When using LTO with GCC 10 or newer, this enables the `-flto=auto`
flag to parallelize processing of ltotrans units during the link
process, significantly speeding up compile times. This translates to a
a roughly 7-10% improvement in build times on my own test systems using
GCC 15, though will probably not be as dramatic of an improvement on CI.

This should have no impact on runtime performance of the agent, as the
individual ltotrans units are already processed in isolation from each
other even without this switch being enabled, and thus parallelizing
that process has no impact on the generated code.