Rename default port for the OpenTelemetry Collector
In 2019, OpenTracing and OpenCensus were anounced to be merging. After
this, the OpenCensus code base was
[sunset](https://opentelemetry.io/blog/2023/sunsetting-opencensus/),
including its Prometheus exporter. Since the default port of the
Prometheus exporter in OpenTelemetry Collector is the same as this, the
latter is used extensively, and the projects are now merged, we can
safely rename this to `otelcol`, which is the lowercased version of the accepted shorthand [OTelCol](https://opentelemetry.io/docs/concepts/glossary/#otelcol).