Limit IBM plugin builds to 64-bit x86 Linux.
This is the only platform we currently correctly handle the MQ library
dependencies for at the moment. Long term we will almost certainly relax
this and support a wider variety of platforms, but for now this is all
we’re worrying about.

This also splits the platform detection code to it’s own CMake module,
and adds new detection for the CPU architecture to enable this selective
build support.