Correctly split MCP registry update to it’s own workflow. (#21089)
* Correctly split MCP registry update to it’s own workflow.

This needs to trigger when your releases are actually published, not
when we start the release process, so it needs to have it’s own workflow
with a release trigger instead of being part of the release workflow (as
we always manually publish the release on GitHub).

This actually simplifies a couple of parts of the job as well, since the
`release` event defaults to using the commit of the release that
triggered the workflow when checking out the repository.

* Address review comments.

(cherry picked from commit 81acfc85b582c1fd78ade45f37fb60902514457c)