Convert RPM specfile to use sysusers for Netdata user creation.
This also splits the user account handling to a dedicated package, which
both simplifies the dependency tree, and also eliminates most of our
preinstall hooks (almost all of them were just creating the accounts if
they didn’t exist).

On RPM 4.19 and newer, simply packaging the sysusers file results in
the resultant RPM auto-generating the user information without any need
for us to supply an install script to generate the accounts.

On older RPM versions, we still ship and install the file, but we also
include a postinstall script that then uses it to create the users (or
falls back to useradd/groupadd).

Use of the RPM integration is keyed off of the presence of the
`add_sysuser` macro that RPM versions with this support provide, instead
of keying off of specific distros and releases, which should avoid the
need for us to update this as new releases of distros we support come
out that use a new enough RPM version.