Update DEB packages to use sysusers for account creation.
Like with the RPMs, this splits the user account handling to it’s own
package, which simplifies both our set of maintainer scripts and our
overall dependency tree.

Unlike RPM, dpkg does not currently have support for automatic user
account handling (though there is a draft proposal in the works to add
this), so we need to always have appropriate maintainer scripts to
create the users.