Switch static builds and local builds to use sysusers when available.
Also, centralize the user account handling code between the local and
static builds.

The updated code will directly install the sysusers.d file if the system
appears to support sysusers.d user generation, and will then use
systemd-sysusers for account creation. In the event that this is not
supported, it falls back to the old behavior.