Trim list of supplementary groups for new installs.
Almost all of the supplementary groups we currently add the Netdata
user to were used by the older log-based collectors written in Python.
On modern installs where data collection is done by the Go plugin out of
the box, we don’t need these group memberships due to the use of file
capabilities, so we shouldn’t be adding the Netdata user to them.