Remove DEB dashboard postinst/postrm hooks.
They haven’t actually been included in the package since we switched to
using a separate package, and nothing broke, so we obviously don’t
actually need them. What they were doing was questionable anyway
(dpkg-statoverride is supposed to be for the administrator to adjust
things, not for individual packages to fix their supposedly broken
permissions).