Switch to git-cliff for changelog generation.
This should completely eliminate the reliability issues in changelog
generation for nightly builds and releases, as git-cliff only inspects
Git history and does not rely on GitHub’s REST API. Note that this _does
not_ require us to switch to using Conventional Commits for our commit
messages (git-cliff provides special additional handling for commits
with messages conforming to this format, but works acceptably well
without that additional handling).

This also makes it extremely easy to generate a changelog locally if
needed, and significantly improves performance of the changelog
generation process (it should now take less than a minute instead of
almost 10 minutes).

The new changelog format is similar, but not identical, to the old
format, but properly tracks changes in a way that doesn’t result in a
tiny and largely useless file.

This involves the following significant changes to the generated
changelogs:
- No authorship information on the changelog entries. If we decide this
  is desirable we can add this, but not doing it means we don’t have to
  rely on GitHub’s APIs, which seem to be the major performance and
  reliability bottleneck with our current changelog generation.
- A given branch’s changelog will now only track the changes and tags in
  that branch. This means that the main branch will not include commits
  for patch releases in it’s changelog, and the release branches will
  not include changes from the main branch since the release branch was
  created that have not made it into a patch release on that branch.

As part of this, the changelog for versions prior to version 2.0.0 has
been split to it’s own file. This helps limit the overall size of the
changelog file (together, it’s over 1 MiB in size), and also speeds up
the generation process as it limits how many commits need to be
processed.